refactor(context): integrate visualization into generate command
- Move visualization from separate command to --visualize flags on context generate
- Add --visualize, --visualize-only, --chart-type, --show-details flags
- Support combined workflows: generate + analyze + visualize in single command
- Improve user experience with natural flag progression: --analyze → --visualize
- Remove separate visualize command to eliminate code duplication
- Maintain all visualization features: bar/pie/tree charts, model comparison, CSV output
- Update help text and examples to reflect integrated approach
- Fix linter error: use template literal instead of string concatenation

Better design: Users now get context → analysis → visualization in natural workflow
instead of requiring separate commands. Follows Unix philosophy while providing
integrated experience similar to git's subcommand model.
