+++ src/commands/context/index.ts
@@ -9,7 +9,6 @@
 import { Command } from "commander";
 import { createSuggestRulesCommand } from "./suggest-rules";
 import { createGenerateCommand } from "./generate";
-import { createVisualizeCommand } from "./visualize";
 
 /**
  * Create the main context command
@@ -22,8 +21,7 @@ export function createContextCommand(): Command {
       `
 Context Management:
   suggest-rules    Get AI-powered rule suggestions for your current task
-  generate         Generate AI context using modular components with optional analysis
-  visualize        Generate visual representation of context token usage
+  generate         Generate AI context using modular components with optional analysis and visualization
 
 Examples:
   minsky context suggest-rules "I need to fix a bug"
@@ -33,10 +31,12 @@ Examples:
   minsky context generate --compare-models gpt-4,claude-3-5-sonnet  # Cross-model comparison
   minsky context generate --output /tmp/test-context.txt
 
-  minsky context visualize                           # Basic bar chart visualization
-  minsky context visualize --chart-type pie         # Pie chart of token distribution
-  minsky context visualize --compare-models gpt-4,claude-3-5-sonnet
-  minsky context visualize --json                   # JSON output for processing
+  # Visualization examples (integrated into generate command)
+  minsky context generate --visualize                  # Generate context with bar chart
+  minsky context generate --visualize-only --chart-type pie  # Only show pie chart
+  minsky context generate --visualize --show-details   # Visualization with detailed breakdown
+  minsky context generate --compare-models gpt-4,claude-3-5-sonnet --visualize  # Compare with charts
+  minsky context generate --csv                        # CSV output for spreadsheets
 
 The context command helps you understand and optimize the information
 available to AI assistants for better collaboration.
@@ -46,7 +46,6 @@ available to AI assistants for better collaboration.
   // Add subcommands
   contextCmd.addCommand(createSuggestRulesCommand());
   contextCmd.addCommand(createGenerateCommand());
-  contextCmd.addCommand(createVisualizeCommand());
 
   return contextCmd;
 }
